Moved eventbus registering/unregistering/onEvent-ing to MainActivity instead of in the fragment. Setting it in the fragment for some reason was causing onEvent to be called twice, thus creating two DetailActivities stacked on top of another. You had to press back twice to get to the grid.
Also updated the DetailFragment to load the movie poster into the Collapsing toolbar (although I'll need to do some adjusting for the zoom level of the image in the toolbar to make sense for tablet layouts).

I need to
1 - add an empty state for the detail fragment or find a way to initalize a default movie in the detail fragment
2 - make the collapsingToolbar poster image look nice and not so damn zoomed in
3 - add the rest of the stuff (like description) to the detail fragment
