+++ app/src/main/res/layout/activity_main.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/container"
     android:layout_width="match_parent"
@@ -7,4 +7,5 @@
     android:background="@color/colorAccentDark"
     tools:context=".MainActivity"
     tools:ignore="MergeRootFrame"
-    tools:layout="@layout/fragment_movie_grid"/>
+    tools:layout="@layout/fragment_movie_grid"
+    class = "com.example.benmedcalf.popularmovies.MoviesGridFragment"/>
