+++ app/src/main/res/layout/empty_state.xml
@@ -4,7 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/empty_state">
-
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
